+++ internal/config/config.go
@@ -1,4 +1,4 @@
-package util
+package config
 
 import (
 	"github.com/spf13/viper"
