+++ pkg/handler/business-buygoods_test.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"testing"
 
-	"github.com/silaselisha/go-daraja/pkg/handler"
 	"github.com/stretchr/testify/require"
 )
 
@@ -39,7 +38,7 @@ func TestBusinessBuyGoods(t *testing.T) {
 				require.NotEmpty(t, buff)
 				require.NoError(t, err)
 
-				var payload handler.BusinessResParams
+				var payload BusinessResParams
 				err = json.Unmarshal(buff, &payload)
 				require.NoError(t, err)
 				require.Equal(t, "0", payload.ResponseCode)
@@ -50,7 +49,7 @@ func TestBusinessBuyGoods(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			client, err := handler.NewDarajaClient("./../../../example")
+			client, err := NewDarajaClient("./../../example")
 			require.NoError(t, err)
 			auth, err := client.ClientAuth()
 			require.NoError(t, err)
