+++ pkg/handler/business-consumer_test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/silaselisha/go-daraja/pkg/handler"
 	"github.com/stretchr/testify/require"
 )
 
@@ -32,7 +31,7 @@ func TestBusinessConsumer(t *testing.T) {
 				require.NoError(t, err)
 				require.NotNil(t, buff)
 
-				var payload handler.BusinessResParams
+				var payload BusinessResParams
 				err = json.Unmarshal(buff, &payload)
 				require.NoError(t, err)
 				require.Equal(t, "0", payload.ResponseCode)
@@ -43,7 +42,7 @@ func TestBusinessConsumer(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			client, err := handler.NewDarajaClient("./../../../example")
+			client, err := NewDarajaClient("./../../example")
 			require.NoError(t, err)
 
 			auth, err := client.ClientAuth()
